Solve the logout problem
