+++ src/Redux/logIn/logInActions.js
@@ -1,9 +1,16 @@
 export const LOGIN_USER_SUCCESS = 'LOGIN_USER_SUCCESS';
+export const LOGOUT_USER_SUCCESS = 'LOGOUT_USER_SUCCESS';
 
 const baseURL = 'http://127.0.0.1:3000/login';
+const logoutURL = 'http://127.0.0.1:3000/logout';
 
 const logInUserSuccess = (users) => ({
   type: LOGIN_USER_SUCCESS,
+  payload: users,
+});
+
+const logOutUserSuccess = (users) => ({
+  type: LOGOUT_USER_SUCCESS,
   payload: users,
 });
 
@@ -28,7 +35,27 @@ export const logInUser = (user) => async (dispatch) => {
   const xyz = await signInUser(user);
   const logInDetails = {
     loggedOut: xyz.logged_out,
-    loggedIn: !xyz.logged_out,
   };
   dispatch(logInUserSuccess(logInDetails));
+};
+
+export const quitUser = async () => {
+  const currUser = {
+    method: 'POST',
+    headers: { 'Content-Type': 'application/json' },
+  };
+  const response = await fetch(logoutURL, currUser);
+  const data = await response.json();
+  localStorage.setItem('ANYNAME', JSON.stringify(data.logged_out));
+  // console.log('logout data', data);
+  return data;
+};
+
+export const logOutUser = () => async (dispatch) => {
+  const xyz = await quitUser();
+  // console.log('xyz', xyz);
+  const logOutDetails = {
+    loggedOut: xyz.logged_out,
+  };
+  dispatch(logOutUserSuccess(logOutDetails));
 };
