+++ src/Redux/rootReducer.js
@@ -1,12 +1,10 @@
 import { combineReducers } from 'redux';
 import signupReducer from './Signup/signupReducer';
 import logInReducer from './logIn/logInReducer';
-import logOutReducer from './logOut/logOutReducer';
 
 const rootReducer = combineReducers({
   signupReducer,
   logInReducer,
-  logOutReducer,
 });
 
 export default rootReducer;
